Moved ActionStore to its own file
