+++ src/blockMonitor/blockchainMachine.ts
@@ -1,72 +1,9 @@
 import { BlockProcessor } from "./blockProcessor";
 import { IBlockStub, StartStopService, ApplicationError } from "../dataEntities";
 import { Component, AnchorState, ComponentAction } from "./component";
+import { ActionStore } from "./actionStore";
 import { BlockItemStore } from "../dataEntities/block";
 import { Lock } from "../utils/lock";
-
-import { LevelUp } from "levelup";
-import EncodingDown from "encoding-down";
-const sub = require("subleveldown");
-import uuid = require("uuid/v4");
-
-export interface ActionAndId {
-    id: string;
-    action: ComponentAction;
-}
-
-export class ActionStore extends StartStopService {
-    private readonly subDb: LevelUp<EncodingDown<string, any>>;
-    private actions: Map<string, Set<ActionAndId>> = new Map();
-
-    constructor(db: LevelUp<EncodingDown<string, any>>) {
-        super("action-store");
-        this.subDb = sub(db, `action-store`, { valueEncoding: "json" });
-    }
-
-    protected async startInternal() {
-        // load existing actions from the db
-        for await (const record of this.subDb.createReadStream()) {
-            const { key, value } = (record as any) as { key: string; value: ComponentAction };
-
-            const i = key.indexOf(":");
-            const componentName = key.substring(0, i);
-            const actionId = key.substring(i + 1);
-
-            const actionWithId = { id: actionId, action: value };
-
-            const componentActions = this.actions.get(componentName);
-            if (componentActions) componentActions.add(actionWithId);
-            else this.actions.set(componentName, new Set([actionWithId]));
-        }
-    }
-    protected async stopInternal() {}
-
-    public getActions(componentName: string) {
-        return this.actions.get(componentName) || new Set();
-    }
-
-    public async storeActions(componentName: string, actions: ComponentAction[]) {
-        // we forge unique ids for actions to uniquely distinguish them in the db
-        const actionsWithId = actions.map(a => ({ id: uuid(), action: a }));
-
-        const componentSet = this.actions.get(componentName);
-        if (componentSet) actionsWithId.forEach(a => componentSet.add(a));
-        else this.actions.set(componentName, new Set(actionsWithId));
-
-        let batch = this.subDb.batch();
-        actionsWithId.forEach(actionWithId => {
-            batch = batch.put(componentName + ":" + actionWithId.id, actionWithId.action);
-        });
-        await batch.write();
-    }
-
-    public async removeAction(componentName: string, actionAndId: ActionAndId) {
-        const actions = this.actions.get(componentName);
-        if (!actions) return;
-        else actions.delete(actionAndId);
-        await this.subDb.del(componentName + ":" + actionAndId.id);
-    }
-}
 
 /**
  * Generic class to handle the anchor state of a blockchain state machine.
