+++ src/service.ts
@@ -12,7 +12,8 @@ import { LevelUp } from "levelup";
 import encodingDown from "encoding-down";
 import { blockFactory } from "./blockMonitor";
 import { Block, BlockItemStore } from "./dataEntities/block";
-import { BlockchainMachine, ActionStore } from "./blockMonitor/blockchainMachine";
+import { BlockchainMachine } from "./blockMonitor/blockchainMachine";
+import { ActionStore } from "./blockMonitor/actionStore";
 import { Logger } from "./logger";
 import path from "path";
 import { GasQueue } from "./responder/gasQueue";
